✨ Add logging level option.
