+++ src/cli/mod.rs
@@ -1,4 +1,5 @@
 pub mod checkout;
 pub mod commit;
 pub mod context;
+pub mod log;
 pub mod templates;
